cloud_io: adopt cloud_storage::{upload/download}_result
I intend for the generic code to continue to use these as their
interface, in order to avoid some major code churn.
